+++ bencher/__init__.py
@@ -3,6 +3,7 @@ import warnings
 warnings.filterwarnings("ignore", message="Unable to import Axes3D", category=UserWarning)
 
 from .bencher import Bench, BenchCfg, BenchRunCfg
+from .bench_cfg import ShowMode
 from .bench_runner import BenchRunner
 from .example.benchmark_data import ExampleBenchCfg
 from .bench_plot_server import BenchPlotServer