+++ bencher/example/example_image.py
@@ -71,4 +71,4 @@ def example_image(run_cfg: bn.BenchRunCfg | None = None) -> bn.Bench:
 
 
 if __name__ == "__main__":
-    bn.run(example_image, level=3)
+    bn.run(example_image, level=3, show=bn.ShowMode.HTML)